requirements.txt
selenium==4.22.0
pytest==8.2.2
Faker==26.1.0
allure-pytest==2.13.2
pytest-xdist==3.3.1
